Added support for the Microsoft Visual C/C++ 2010 compiler.
git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@105 5fe0b3f3-48df-4727-b460-e6dda9226989

+++ Libs/Bitmap/Bitmap.cpp
@@ -48,7 +48,7 @@ void jpeg_FileSys_src(j_decompress_ptr cinfo, cf::FileSys::InFileI* JpegFile_);
 static void MyFileSysBasedPngReadFunction(png_structp png_ptr, png_bytep data, png_size_t length)
 {
     cf::FileSys::InFileI* PngFile  =static_cast<cf::FileSys::InFileI*>(png_get_io_ptr(png_ptr));
-    const size_t          BytesRead=PngFile->Read((char*)data, length);
+    const uint32_t        BytesRead=PngFile->Read((char*)data, uint32_t(length));
 
     if (BytesRead<length) png_error(png_ptr, "Read Error");
 }
